merge

+++ GAC-EJBFacade/ejbModule/remote/FActivityServicesRemote.java
@@ -8,7 +8,7 @@ import model.Activity;
 public interface FActivityServicesRemote {
 	 public void addItem(Activity i);
 
-	    public Activity findItem(long id) ;
+	    public Activity findItem(Integer id) ;
 
 	    public void deleteItem(Activity i) ;
