+++ GAC-EJBFacade/ejbModule/remote/FMessageServicesRemote.java
@@ -8,7 +8,7 @@ import model.Message;
 public interface FMessageServicesRemote {
 	 public void addItem(Message i);
 
-	    public Message findItem(long id) ;
+	    public Message findItem(Integer id) ;
 
 	    public void deleteItem(Message i) ;
